Refactor and reorganize queries and other code
