+++ cypress/tsconfig.json
@@ -1,8 +1,9 @@
 {
   "compilerOptions": {
+    "module": "esnext",
     "esModuleInterop": true,
     "target": "es5",
-    "lib": ["es5", "dom"],
+    "lib": ["es5", "dom", "esnext"],
     "types": ["cypress", "node"]
   },
   "include": ["**/*.ts"]
