+++ src/hooks/data/queries/index.ts
@@ -1,0 +1,13 @@
+export {
+  accommodationQuery,
+  readPersonAccommodationsQuery,
+  searchAccommodationsQuery,
+} from './accommodations'
+export { readCommunityMembersQuery, readCommunityQuery } from './community'
+export { contactRequestsQuery, contactsQuery } from './contacts'
+export {
+  hospexDocumentQuery,
+  privateProfileAndHospexDocumentQuery,
+} from './hospex'
+export { inboxMessagesQuery, messages, threads } from './messages'
+export { profileDocuments, webIdProfileQuery } from './profile'
